There is still a bug where if you refund the pips while a temp is decaying it crashes the unity GUI.
Need to fix that and have it add extra pips and remove them when they expire.

+++ DarwinsDescent/Assets/Scripts/Characters/MonoBehaviour/Pips/HPPipModel.cs
@@ -20,7 +20,7 @@ namespace DarwinsDescent
         }
         public state CurState;
 
-        public int TempTime;
+        public float TempTime;
 
         public Image PipDisplayImage;
     }
